Scraper/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
supabase==2.11.0
python-dotenv==1.0.1
lxml==5.1.0
httpx==0.27.2
websockets>=13.0
